Ignore old cookies and try to keep a session active
The login cookie usually expires at the end of the browser session.
If the browser is not restarted, that can be a long time.

We'd like to ignore sessions older than X hours, but prevent an active
session from being destroyed.

What this does is, it completely disregards a session cookie if it was
created more than 12 hours ago.

To prevent this happening as a user browses the site, we provide a
session extension mechanism.

If the current *valid* session cookie was created more than 30 minutes
ago (and obviously less than 12 hours ago, since it's valid) - it's a
candidate for regeneration and is marked via an attribute on the request
object.

An additional after() listener takes care to perform the actual extension
after the response is created.